Add files via upload
This python code will generate a gui that will allow you to take videos and images.  It will save them as alpha image or video with a time stamp.  The video/images are also annotated, so you can change or eliminate that in the code.  You should make it executable, so chmod +x script_name.py then you can just double click on it to execute the gui.  I am using imagej on the pi, but you can use a python script to count particles.  I do intend to add this to the code soon (google has lots of examples).